feat: add query params to visualize to show directly a node from case
