+++ components/dashboard/sidebar.tsx
@@ -11,7 +11,7 @@ import {
   LogOutIcon,
   FlaskConicalIcon,
   PencilIcon,
-  EyeIcon,
+  NetworkIcon,
   GlobeIcon,
   ConstructionIcon
 } from 'lucide-react'
@@ -131,7 +131,7 @@ const Sidebar = () => {
                 className='flex items-center gap-3 rounded-lg px-3 py-2 text-gray-500 transition-all hover:text-gray-900 dark:text-gray-400 dark:hover:text-gray-50'
                 href='/dashboard/visualize'
               >
-                <EyeIcon className='h-4 w-4' />
+                <NetworkIcon className='h-4 w-4 -rotate-90' />
                 Visualize
               </Link>
             </div>
